Use better inheritance
