+++ src/backend/app/Models/Transaction/AgentTransaction.php
@@ -3,23 +3,19 @@
 namespace App\Models\Transaction;
 
 use App\Models\Agent;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Support\Carbon;
 
 /**
  * @property      int                                   $id
  * @property      int                                   $agent_id
  * @property      string                                $mobile_device_id
- * @property      int                                   $status
  * @property      Carbon|null                           $created_at
  * @property      Carbon|null                           $updated_at
  * @property      string|null                           $manufacturer_transaction_type
  * @property      int|null                              $manufacturer_transaction_id
  * @property-read Agent|null                            $agent
- * @property-read Collection<int, TransactionConflicts> $conflicts
  * @property-read Model|null                            $manufacturerTransaction
  * @property-read Transaction|null                      $transaction
  */
@@ -31,13 +27,6 @@ class AgentTransaction extends BasePaymentProviderTransaction {
      */
     public function agent(): BelongsTo {
         return $this->belongsTo(Agent::class);
-    }
-
-    /**
-     * @return MorphMany<TransactionConflicts, $this>
-     */
-    public function conflicts(): MorphMany {
-        return $this->morphMany(TransactionConflicts::class, 'transaction');
     }
 
     public static function getTransactionName(): string {
