+++ src/backend/app/Models/Transaction/BasePaymentProviderTransaction.php
@@ -4,7 +4,9 @@ namespace App\Models\Transaction;
 
 use App\Models\Base\BaseModel;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Eloquent\Relations\MorphOne;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
 
@@ -12,8 +14,20 @@ use Illuminate\Database\Eloquent\Relations\MorphTo;
  * Base class for payment provider transaction models.
  *
  * This abstract class provides common relationship methods that are shared across all payment provider transaction implementations.
+ *
+ * @property      int                                   $status
+ * @property-read Collection<int, TransactionConflicts> $conflicts
  */
 abstract class BasePaymentProviderTransaction extends BaseModel {
+    /**
+     * Get the conflicts detected for this transaction.
+     *
+     * @return MorphMany<TransactionConflicts, $this>
+     */
+    public function conflicts(): MorphMany {
+        return $this->morphMany(TransactionConflicts::class, 'transaction');
+    }
+
     /**
      * Get the base transaction relationship.
      *
