+++ src/backend/app/Models/Transaction/CashTransaction.php
@@ -3,21 +3,17 @@
 namespace App\Models\Transaction;
 
 use App\Models\User;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Support\Carbon;
 
 /**
  * @property      int                                   $id
  * @property      int                                   $user_id
- * @property      int                                   $status
  * @property      string|null                           $manufacturer_transaction_type
  * @property      int|null                              $manufacturer_transaction_id
  * @property      Carbon|null                           $created_at
  * @property      Carbon|null                           $updated_at
- * @property-read Collection<int, TransactionConflicts> $conflicts
  * @property-read Model|null                            $manufacturerTransaction
  * @property-read Transaction|null                      $transaction
  * @property-read User|null                             $user
@@ -30,13 +26,6 @@ class CashTransaction extends BasePaymentProviderTransaction {
      */
     public function user(): BelongsTo {
         return $this->belongsTo(User::class);
-    }
-
-    /**
-     * @return MorphMany<TransactionConflicts, $this>
-     */
-    public function conflicts(): MorphMany {
-        return $this->morphMany(TransactionConflicts::class, 'transaction');
     }
 
     public static function getTransactionName(): string {
