+++ src/backend/app/Plugins/PesapalPaymentProvider/Services/PesapalTransactionService.php
@@ -8,26 +8,22 @@ use App\Jobs\ProcessPayment;
 use App\Models\Address\Address;
 use App\Models\Meter\Meter;
 use App\Models\SolarHomeSystem;
+use App\Models\Transaction\BasePaymentProviderTransaction;
 use App\Models\Transaction\Transaction;
 use App\Plugins\PesapalPaymentProvider\Models\PesapalTransaction;
 use App\Plugins\PesapalPaymentProvider\Modules\Api\PesapalApiService;
 use App\Plugins\PesapalPaymentProvider\Modules\Api\Resources\GetTransactionStatusResource;
-use App\Services\AbstractPaymentAggregatorTransactionService;
-use App\Services\DeviceService;
-use App\Services\Interfaces\IBaseService;
-use App\Services\Interfaces\PaymentInitiator;
-use App\Services\PersonService;
-use Illuminate\Database\Eloquent\Collection;
-use Illuminate\Pagination\LengthAwarePaginator;
+use App\Services\AbstractRedirectPaymentInitiatorService;
 use Illuminate\Support\Facades\DB;
 use Ramsey\Uuid\Uuid;
 
 /**
- * @implements IBaseService<PesapalTransaction>
+ * @extends AbstractRedirectPaymentInitiatorService<PesapalTransaction>
  */
-class PesapalTransactionService extends AbstractPaymentAggregatorTransactionService implements IBaseService, PaymentInitiator {
+class PesapalTransactionService extends AbstractRedirectPaymentInitiatorService {
     public function __construct(
         private Meter $meter,
+        private SolarHomeSystem $solarHomeSystem,
         private Address $address,
         private Transaction $transaction,
         private PesapalTransaction $pesapalTransaction,
@@ -36,6 +32,7 @@ class PesapalTransactionService extends AbstractPaymentAggregatorTransactionServ
     ) {
         parent::__construct(
             $this->meter,
+            $this->solarHomeSystem,
             $this->address,
             $this->transaction,
             $this->pesapalTransaction
@@ -46,20 +43,17 @@ class PesapalTransactionService extends AbstractPaymentAggregatorTransactionServ
      * @return array<string, mixed>
      */
     public function initializeTransactionData(): array {
-        $orderId = Uuid::uuid4()->toString();
-        $referenceId = Uuid::uuid4()->toString();
-
         return [
-            'order_id' => $orderId,
-            'reference_id' => $referenceId,
-            'serial_id' => $this->getSerialId(),
+            'order_id' => Uuid::uuid4()->toString(),
+            'reference_id' => Uuid::uuid4()->toString(),
+            'serial_id' => $this->meterSerialNumber,
             'status' => PesapalTransaction::STATUS_REQUESTED,
             'currency' => $this->credentialService->getCredentials()->getCurrency(),
-            'customer_id' => $this->getCustomerId(),
-            'amount' => $this->getAmount(),
+            'customer_id' => $this->customerId,
+            'amount' => $this->amount,
             'metadata' => [
-                'serial_id' => $this->getSerialId(),
-                'customer_id' => $this->getCustomerId(),
+                'serial_id' => $this->meterSerialNumber,
+                'customer_id' => $this->customerId,
             ],
         ];
     }
@@ -82,24 +76,6 @@ class PesapalTransactionService extends AbstractPaymentAggregatorTransactionServ
 
     public function getByReferenceId(string $referenceId): ?PesapalTransaction {
         return $this->pesapalTransaction->newQuery()->where('reference_id', '=', $referenceId)->first();
-    }
-
-    /**
-     * @return Collection<int, PesapalTransaction>
-     */
-    public function getByStatus(int $status): Collection {
-        return $this->pesapalTransaction->newQuery()->where('status', '=', $status)->get();
-    }
-
-    public function getById(int $id): ?PesapalTransaction {
-        return $this->pesapalTransaction->newQuery()->find($id);
-    }
-
-    public function update($pesapalTransaction, array $pesapalTransactionData): PesapalTransaction {
-        $pesapalTransaction->update($pesapalTransactionData);
-        $pesapalTransaction->fresh();
-
-        return $pesapalTransaction;
     }
 
     public function create(array $pesapalTransactionData): PesapalTransaction {
@@ -128,26 +104,8 @@ class PesapalTransactionService extends AbstractPaymentAggregatorTransactionServ
         }
     }
 
-    public function delete($pesapalTransaction): ?bool {
-        return $pesapalTransaction->delete();
-    }
-
-    public function getAll(?int $limit = null): Collection|LengthAwarePaginator {
-        $query = $this->pesapalTransaction->newQuery();
-
-        if ($limit) {
-            return $query->paginate($limit);
-        }
-
-        return $this->pesapalTransaction->newQuery()->get();
-    }
-
     public function getPesapalTransaction(): PesapalTransaction {
         return $this->getPaymentAggregatorTransaction();
-    }
-
-    public function getSerialId(): ?string {
-        return $this->getMeterSerialNumber();
     }
 
     public function processSuccessfulPayment(int $companyId, PesapalTransaction $transaction): void {
@@ -157,6 +115,11 @@ class PesapalTransactionService extends AbstractPaymentAggregatorTransactionServ
         $transaction->save();
     }
 
+    /**
+     * TODO: Unlike PaystackTransactionService::processFailedPayment, this does not mark the
+     * related base Transaction as failed. The two diverged during copy-paste; reconcile once
+     * there is test coverage confirming the intended behaviour.
+     */
     public function processFailedPayment(PesapalTransaction $transaction): void {
         $transaction->setStatus(PesapalTransaction::STATUS_FAILED);
         $transaction->save();
@@ -186,6 +149,34 @@ class PesapalTransactionService extends AbstractPaymentAggregatorTransactionServ
         return $status;
     }
 
+    protected function resolveCurrency(): string {
+        $credential = $this->credentialService->getCredentials();
+
+        return $credential->getCurrency() ?: config('pesapal-payment-provider.currency.default', 'KES');
+    }
+
+    protected function requestedStatus(): int {
+        return PesapalTransaction::STATUS_REQUESTED;
+    }
+
+    /**
+     * @param PesapalTransaction $transaction
+     *
+     * @return array<string, mixed>
+     */
+    protected function submitToProvider(BasePaymentProviderTransaction $transaction, string $sender): array {
+        $result = $this->pesapalApiService->submitOrder($transaction, $sender);
+        if ($result['error']) {
+            throw new \RuntimeException('Pesapal initialization failed: '.$result['error']);
+        }
+
+        return [
+            'redirect_url' => $result['redirect_url'],
+            'order_tracking_id' => $result['order_tracking_id'],
+            'merchant_reference' => $result['merchant_reference'],
+        ];
+    }
+
     private function applyStatusCode(PesapalTransaction $transaction, ?int $statusCode, int $companyId): void {
         switch ($statusCode) {
             case GetTransactionStatusResource::STATUS_COMPLETED:
@@ -199,165 +190,6 @@ class PesapalTransactionService extends AbstractPaymentAggregatorTransactionServ
                 $transaction->setStatus(PesapalTransaction::STATUS_ABANDONED);
                 $transaction->save();
                 break;
-        }
-    }
-
-    /**
-     * @return array{transaction: Transaction, provider_data: array<string, mixed>}
-     */
-    public function initiatePayment(
-        float $amount,
-        string $sender,
-        string $message,
-        string $type,
-        int $customerId,
-        ?string $serialId = null,
-    ): array {
-        $deviceType = null;
-        if ($serialId !== null) {
-            $device = app(DeviceService::class)->getBySerialNumber($serialId);
-            $deviceType = $device?->device_type;
-        }
-
-        $credential = $this->credentialService->getCredentials();
-
-        try {
-            DB::connection('tenant')->beginTransaction();
-
-            $pesapalTxn = $this->pesapalTransaction->newQuery()->create([
-                'amount' => $amount,
-                'currency' => $credential->getCurrency() ?: config('pesapal-payment-provider.currency.default', 'KES'),
-                'order_id' => Uuid::uuid4()->toString(),
-                'reference_id' => Uuid::uuid4()->toString(),
-                'status' => PesapalTransaction::STATUS_REQUESTED,
-                'customer_id' => $customerId,
-                'serial_id' => $serialId,
-                'device_type' => $deviceType,
-                'metadata' => ['customer_id' => $customerId, 'serial_id' => $serialId, 'transaction_type' => $type],
-            ]);
-
-            /** @var Transaction $transaction */
-            $transaction = $pesapalTxn->transaction()->create([
-                'amount' => $amount,
-                'sender' => $sender,
-                'message' => $message,
-                'type' => $type,
-            ]);
-
-            $result = $this->pesapalApiService->submitOrder($pesapalTxn, $sender);
-            if ($result['error']) {
-                throw new \RuntimeException('Pesapal initialization failed: '.$result['error']);
-            }
-
-            DB::connection('tenant')->commit();
-
-            return [
-                'transaction' => $transaction,
-                'provider_data' => [
-                    'redirect_url' => $result['redirect_url'],
-                    'order_tracking_id' => $result['order_tracking_id'],
-                    'merchant_reference' => $result['merchant_reference'],
-                ],
-            ];
-        } catch (\Exception $e) {
-            DB::connection('tenant')->rollBack();
-            throw $e;
-        }
-    }
-
-    /**
-     * @param array<string, mixed> $transactionData
-     */
-    public function createPublicTransaction(array $transactionData): PesapalTransaction {
-        $transactionData['status'] = PesapalTransaction::STATUS_REQUESTED;
-        $transactionData['metadata'] = [
-            'serial_id' => $transactionData['serial_id'],
-            'customer_id' => $transactionData['customer_id'],
-            'public_payment' => true,
-        ];
-
-        if (isset($transactionData['agent_id']) && $transactionData['agent_id']) {
-            $transactionData['metadata']['agent_id'] = $transactionData['agent_id'];
-        }
-
-        return $this->pesapalTransaction->newQuery()->create($transactionData);
-    }
-
-    public function validateMeterSerial(string $serialId): bool {
-        $meter = $this->meter->newQuery()
-            ->where('serial_number', $serialId)
-            ->where('in_use', 1)
-            ->first();
-
-        return $meter !== null;
-    }
-
-    public function validateSHSSerial(string $serialId): bool {
-        $shs = app()->make(SolarHomeSystem::class)
-            ->newQuery()
-            ->where('serial_number', $serialId)
-            ->first();
-
-        return $shs !== null;
-    }
-
-    public function validateDeviceSerial(string $serialId, string $deviceType = 'meter'): bool {
-        if ($deviceType === 'shs') {
-            return $this->validateSHSSerial($serialId);
-        }
-
-        return $this->validateMeterSerial($serialId);
-    }
-
-    public function validatePaymentOwner(string $serialId, float $amount): void {
-        if (!$this->validateMeterSerial($serialId)) {
-            throw new \Exception('Invalid meter serial number');
-        }
-
-        if ($amount <= 0) {
-            throw new \Exception('Invalid payment amount');
-        }
-    }
-
-    public function getCustomerIdByMeterSerial(string $serialId): ?int {
-        $meter = $this->meter->newQuery()
-            ->where('serial_number', $serialId)
-            ->where('in_use', 1)
-            ->first();
-
-        if (!$meter) {
-            return null;
-        }
-
-        return $meter->device->person->id;
-    }
-
-    public function getCustomerIdBySHSSerial(string $serialId): ?int {
-        $shs = app()->make(SolarHomeSystem::class)
-            ->newQuery()
-            ->where('serial_number', $serialId)
-            ->first();
-
-        if (!$shs) {
-            return null;
-        }
-
-        $device = $shs->device()->first();
-        if (!$device || !$device->person) {
-            return null;
-        }
-
-        return (int) $device->person->id;
-    }
-
-    public function getCustomerPhoneByCustomerId(int $customerId): ?string {
-        try {
-            $personService = app()->make(PersonService::class);
-            $person = $personService->getById($customerId);
-
-            return (string) $person->addresses->first()->phone;
-        } catch (\Exception) {
-            return null;
         }
     }
 }
