+++ src/backend/app/Plugins/SmsTransactionParser/Models/SmsTransaction.php
@@ -6,10 +6,7 @@ namespace App\Plugins\SmsTransactionParser\Models;
 
 use App\Models\Transaction\BasePaymentProviderTransaction;
 use App\Models\Transaction\Transaction;
-use App\Models\Transaction\TransactionConflicts;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Support\Carbon;
 
 /**
@@ -20,12 +17,10 @@ use Illuminate\Support\Carbon;
  * @property      string                                $sender_phone
  * @property      string|null                           $device_serial
  * @property      string                                $raw_message
- * @property      int                                   $status
  * @property      string|null                           $manufacturer_transaction_type
  * @property      int|null                              $manufacturer_transaction_id
  * @property      Carbon|null                           $created_at
  * @property      Carbon|null                           $updated_at
- * @property-read Collection<int, TransactionConflicts> $conflicts
  * @property-read Model|null                            $manufacturerTransaction
  * @property-read Transaction|null                      $transaction
  */
@@ -62,13 +57,6 @@ class SmsTransaction extends BasePaymentProviderTransaction {
 
     public function setStatus(int $status): void {
         $this->status = $status;
-    }
-
-    /**
-     * @return MorphMany<TransactionConflicts, $this>
-     */
-    public function conflicts(): MorphMany {
-        return $this->morphMany(TransactionConflicts::class, 'transaction');
     }
 
     public function getManufacturerTransferType(): ?string {
