+++ src/backend/app/Plugins/SwiftaPaymentProvider/Models/SwiftaTransaction.php
@@ -4,10 +4,7 @@ namespace App\Plugins\SwiftaPaymentProvider\Models;
 
 use App\Models\Transaction\BasePaymentProviderTransaction;
 use App\Models\Transaction\Transaction;
-use App\Models\Transaction\TransactionConflicts;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Support\Carbon;
 
 /**
@@ -15,13 +12,11 @@ use Illuminate\Support\Carbon;
  * @property      string|null                           $transaction_reference
  * @property      string|null                           $manufacturer_transaction_type
  * @property      int|null                              $manufacturer_transaction_id
- * @property      int                                   $status
  * @property      float                                 $amount
  * @property      string                                $cipher
  * @property      string                                $timestamp
  * @property      Carbon|null                           $created_at
  * @property      Carbon|null                           $updated_at
- * @property-read Collection<int, TransactionConflicts> $conflicts
  * @property-read Model|null                            $manufacturerTransaction
  * @property-read Transaction|null                      $transaction
  */
@@ -36,13 +31,6 @@ class SwiftaTransaction extends BasePaymentProviderTransaction {
 
     public function getAmount(): float {
         return $this->amount;
-    }
-
-    /**
-     * @return MorphMany<TransactionConflicts, $this>
-     */
-    public function conflicts(): MorphMany {
-        return $this->morphMany(TransactionConflicts::class, 'transaction');
     }
 
     public static function getTransactionName(): string {
