+++ src/backend/app/Plugins/WaveMoneyPaymentProvider/Models/WaveMoneyTransaction.php
@@ -4,15 +4,11 @@ namespace App\Plugins\WaveMoneyPaymentProvider\Models;
 
 use App\Models\Transaction\BasePaymentProviderTransaction;
 use App\Models\Transaction\Transaction;
-use App\Models\Transaction\TransactionConflicts;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Support\Carbon;
 
 /**
  * @property      int                                   $id
- * @property      int                                   $status
  * @property      float                                 $amount
  * @property      string                                $order_id
  * @property      string                                $reference_id
@@ -25,7 +21,6 @@ use Illuminate\Support\Carbon;
  * @property      Carbon|null                           $updated_at
  * @property      string|null                           $manufacturer_transaction_type
  * @property      int|null                              $manufacturer_transaction_id
- * @property-read Collection<int, TransactionConflicts> $conflicts
  * @property-read Model|null                            $manufacturerTransaction
  * @property-read Transaction|null                      $transaction
  */
@@ -86,13 +81,6 @@ class WaveMoneyTransaction extends BasePaymentProviderTransaction {
 
     public function setAmount(int $amount): void {
         $this->amount = $amount;
-    }
-
-    /**
-     * @return MorphMany<TransactionConflicts, $this>
-     */
-    public function conflicts(): MorphMany {
-        return $this->morphMany(TransactionConflicts::class, 'transaction');
     }
 
     public static function getTransactionName(): string {
